+++ src/public-api.ts
@@ -3,6 +3,8 @@
  */
 
 export { PanelHeadingDirective } from './lib/directives/panel-heading.directive';
+export { PanelHeaderDirective } from './lib/directives/panel-header.directive';
+export { PanelFooterDirective } from './lib/directives/panel-footer.directive';
 export { PanelComponent } from './lib/components/panel/panel.component';
 export { PanelsComponent } from './lib/components/panels/panels.component';
 export { PanelsConfig } from './lib/services/panels-config.service';